airflow/requirements.txt
apache-airflow-providers-google
pyarrow
requests
glom